Modify random voter generation
Modify the distributions from which social and economic preferences are
sampled. Make social preference be sampled from a bivariate normal
distribution with means of 30 and 70 and standard deviations of 10.
Change the economic preference mean shift to a higher value in order to
decrease the quantity of voters with central economic views.